+++ src/qibocal/calibrations/__init__.py
@@ -1,6 +1,6 @@
 from qibocal.calibrations.characterization.allXY import *
 from qibocal.calibrations.characterization.calibrate_qubit_states import *
-from qibocal.calibrations.characterization.cz import *
+from qibocal.calibrations.characterization.chevron import *
 from qibocal.calibrations.characterization.flipping import *
 from qibocal.calibrations.characterization.qubit_spectroscopy import *
 from qibocal.calibrations.characterization.rabi import *
