+++ src/qibocal/plots/__init__.py
@@ -1,6 +1,6 @@
 from qibocal.plots.allXY import *
 from qibocal.plots.calibrate_qubit_states import *
-from qibocal.plots.cz import *
+from qibocal.plots.chevron import *
 from qibocal.plots.flipping import *
 from qibocal.plots.rabi import *
 from qibocal.plots.ramsey import *
